rename activation to activations and rename initialization to initializations
