+++ npdl/optimizers.py
@@ -1,4 +1,3 @@
-
 """
 Functions to generate Theano update dictionaries for training.
 
@@ -15,8 +14,8 @@ Using :class:`SGD` to define an update dictionary for a toy
 example network:
 
 >>> import npdl
->>> from npdl.activation import ReLU
->>> from npdl.activation import Softmax
+>>> from npdl.activations import ReLU
+>>> from npdl.activations import Softmax
 >>> from npdl.objectives import SCCE
 >>> model = npdl.model.Model()
 >>> model.add(npdl.layers.Dense(n_out=100, n_in=50, activation=ReLU()))
